src/requirements.txt
numpy
Pillow
screeninfo
